requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
pillow
opencv-python
tqdm

